reboot02-resume-parser-service/requirements.txt
Flask
gunicorn
pdfplumber
spacy
python-dotenv
google-generativeai
python-dateutil
flask-cors
google-cloud-aiplatform[generative]>=1.5.0